move snat model definitions
