+++ crates/defguard_core/src/enterprise/db/models/mod.rs
@@ -3,3 +3,4 @@ pub mod activity_log_stream;
 pub mod api_tokens;
 pub mod enterprise_settings;
 pub mod openid_provider;
+pub mod snat;
